Signup - fix page structure
