+++ app/onboarding/page.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import OnboardingStep1 from "./onboarding";
+
+export default function OnboardingStepPage() {
+  return <OnboardingStep1 />;
+}
