+++ app/onboarding/style.tsx
